SPCM offsets are 16-bit

+++ project/vgmstream.cpp
@@ -2000,7 +2000,7 @@ void VGMStream::encodeNoteItem(const Project& project, const StreamNoteItem* ite
                     data.append(0xF0 | item->channel() + DPCM_CHANNELS);
                     data.append(att);
 
-                    quint32 offset = rom.offsets()[rcs->keySampleMappings()[item->note().key()]];
+                    quint16 offset = rom.offsets()[rcs->keySampleMappings()[item->note().key()]];
                     quint16 size = rom.size(rcs->keySampleMappings()[item->note().key()]);
 
                     data.append(0xD0 | item->channel() + DPCM_CHANNELS);
